backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
pymysql==1.1.0
cryptography==41.0.7
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
apscheduler==3.10.4
python-dateutil==2.8.2
pytz==2023.3
